Extract "updateMyProfile" GraphQL mutation tests into its own file
The one giant GraphQL test file has grown too big. It's time to split it
into smaller pieces.
