+++ profiles/tests/test_profiles_graphql_api.py
@@ -11,7 +11,6 @@ from helusers.authz import UserAuthorization
 
 from open_city_profile.consts import (
     API_NOT_IMPLEMENTED_ERROR,
-    INVALID_EMAIL_FORMAT_ERROR,
     PERMISSION_DENIED_ERROR,
     PROFILE_DOES_NOT_EXIST_ERROR,
     TOKEN_EXPIRED_ERROR,
@@ -48,755 +47,6 @@ from .factories import (
     TemporaryReadAccessTokenFactory,
     VerifiedPersonalInformationFactory,
 )
-
-
-def test_normal_user_can_update_profile(rf, user_gql_client, email_data, profile_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    email = profile.emails.first()
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                            nickname: \"${nickname}\",
-                            updateEmails:[
-                                {
-                                    id: \"${email_id}\",
-                                    emailType: ${email_type},
-                                    email:\"${email}\",
-                                    primary: ${primary}
-                                }
-                            ]
-                        }
-                    }
-                ) {
-                    profile{
-                        nickname,
-                        emails{
-                            edges{
-                            node{
-                                id
-                                email
-                                emailType
-                                primary
-                                verified
-                            }
-                            }
-                        }
-                    }
-                }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "nickname": profile_data["nickname"],
-                "emails": {
-                    "edges": [
-                        {
-                            "node": {
-                                "id": to_global_id(type="EmailNode", id=email.id),
-                                "email": email_data["email"],
-                                "emailType": email_data["email_type"],
-                                "primary": email_data["primary"],
-                                "verified": False,
-                            }
-                        }
-                    ]
-                },
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        nickname=profile_data["nickname"],
-        email_id=to_global_id(type="EmailNode", id=email.id),
-        email=email_data["email"],
-        email_type=email_data["email_type"],
-        primary=str(email_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_add_email(rf, user_gql_client, email_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    email = profile.emails.first()
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                            addEmails:[
-                                {emailType: ${email_type}, email:\"${email}\", primary: ${primary}}
-                            ]
-                        }
-                    }
-            ) {
-                profile{
-                    emails{
-                        edges{
-                        node{
-                            email
-                            emailType
-                            primary
-                            verified
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "emails": {
-                    "edges": [
-                        {
-                            "node": {
-                                "email": email.email,
-                                "emailType": email.email_type.name,
-                                "primary": email.primary,
-                                "verified": False,
-                            }
-                        },
-                        {
-                            "node": {
-                                "email": email_data["email"],
-                                "emailType": email_data["email_type"],
-                                "primary": not email_data["primary"],
-                                "verified": False,
-                            }
-                        },
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        email=email_data["email"],
-        email_type=email_data["email_type"],
-        primary=str(not email_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_cannot_add_invalid_email(rf, user_gql_client, email_data):
-    ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                            addEmails:[
-                                {emailType: ${email_type}, email:\"${email}\", primary: ${primary}}
-                            ]
-                        }
-                    }
-            ) {
-                profile{
-                    emails{
-                        edges{
-                        node{
-                            id
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    mutation = t.substitute(
-        email="!dsdsd{}{}{}{}{}{",
-        email_type=email_data["email_type"],
-        primary=str(not email_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert "code" in executed["errors"][0]["extensions"]
-    assert executed["errors"][0]["extensions"]["code"] == INVALID_EMAIL_FORMAT_ERROR
-
-
-def test_normal_user_cannot_update_email_to_invalid_format(
-    rf, user_gql_client, email_data
-):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    email = profile.emails.first()
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        updateEmails:[
-                            {
-                                id: \"${email_id}\",
-                                emailType: ${email_type},
-                                email:\"${email}\",
-                                primary: ${primary}
-                            }
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    emails{
-                        edges{
-                        node{
-                            id
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    mutation = t.substitute(
-        email_id=to_global_id(type="EmailNode", id=email.id),
-        email="!dsdsd{}{}{}{}{}{",
-        email_type=email_data["email_type"],
-        primary=str(email_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert "code" in executed["errors"][0]["extensions"]
-    assert executed["errors"][0]["extensions"]["code"] == INVALID_EMAIL_FORMAT_ERROR
-
-
-def test_normal_user_can_add_phone(rf, user_gql_client, phone_data):
-    ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        addPhones:[
-                            {phoneType: ${phone_type}, phone:\"${phone}\", primary: ${primary}}
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    phones{
-                        edges{
-                        node{
-                            phone
-                            phoneType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "phones": {
-                    "edges": [
-                        {
-                            "node": {
-                                "phone": phone_data["phone"],
-                                "phoneType": phone_data["phone_type"],
-                                "primary": phone_data["primary"],
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        phone=phone_data["phone"],
-        phone_type=phone_data["phone_type"],
-        primary=str(phone_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_add_address(rf, user_gql_client, address_data):
-    ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                            addAddresses: [
-                                {
-                                    addressType: ${address_type},
-                                    address:\"${address}\",
-                                    postalCode: \"${postal_code}\",
-                                    city: \"${city}\",
-                                    countryCode: \"${country_code}\",
-                                    primary: ${primary}
-                                }
-                            ]
-                        }
-                    }
-                ) {
-                profile {
-                    addresses {
-                        edges {
-                            node {
-                                address
-                                postalCode
-                                city
-                                countryCode
-                                addressType
-                                primary
-                            }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "addresses": {
-                    "edges": [
-                        {
-                            "node": {
-                                "address": address_data["address"],
-                                "postalCode": address_data["postal_code"],
-                                "city": address_data["city"],
-                                "countryCode": address_data["country_code"],
-                                "addressType": address_data["address_type"],
-                                "primary": address_data["primary"],
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        address=address_data["address"],
-        postal_code=address_data["postal_code"],
-        city=address_data["city"],
-        country_code=address_data["country_code"],
-        address_type=address_data["address_type"],
-        primary=str(address_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_update_address(rf, user_gql_client, address_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    address = AddressFactory(profile=profile)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        updateAddresses:[
-                            {
-                                id: \"${address_id}\",
-                                addressType: ${address_type},
-                                address:\"${address}\",
-                                postalCode:\"${postal_code}\",
-                                city:\"${city}\",
-                                primary: ${primary}
-                            }
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    addresses{
-                        edges{
-                        node{
-                            id
-                            address
-                            postalCode
-                            city
-                            addressType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "addresses": {
-                    "edges": [
-                        {
-                            "node": {
-                                "id": to_global_id(type="AddressNode", id=address.id),
-                                "address": address_data["address"],
-                                "postalCode": address_data["postal_code"],
-                                "city": address_data["city"],
-                                "addressType": address_data["address_type"],
-                                "primary": address_data["primary"],
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        address_id=to_global_id(type="AddressNode", id=address.id),
-        address=address_data["address"],
-        postal_code=address_data["postal_code"],
-        city=address_data["city"],
-        address_type=address_data["address_type"],
-        primary=str(address_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_update_email(rf, user_gql_client, email_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    email = profile.emails.first()
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        updateEmails:[
-                            {
-                                id: \"${email_id}\",
-                                emailType: ${email_type},
-                                email:\"${email}\",
-                                primary: ${primary}
-                            }
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    emails{
-                        edges{
-                        node{
-                            id
-                            email
-                            emailType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "emails": {
-                    "edges": [
-                        {
-                            "node": {
-                                "id": to_global_id(type="EmailNode", id=email.id),
-                                "email": email_data["email"],
-                                "emailType": email_data["email_type"],
-                                "primary": email_data["primary"],
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        email_id=to_global_id(type="EmailNode", id=email.id),
-        email=email_data["email"],
-        email_type=email_data["email_type"],
-        primary=str(email_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_update_phone(rf, user_gql_client, phone_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    phone = PhoneFactory(profile=profile)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        updatePhones:[
-                            {
-                                id: \"${phone_id}\",
-                                phoneType: ${phone_type},
-                                phone:\"${phone}\",
-                                primary: ${primary}
-                            }
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    phones{
-                        edges{
-                        node{
-                            id
-                            phone
-                            phoneType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "phones": {
-                    "edges": [
-                        {
-                            "node": {
-                                "id": to_global_id(type="PhoneNode", id=phone.id),
-                                "phone": phone_data["phone"],
-                                "phoneType": phone_data["phone_type"],
-                                "primary": phone_data["primary"],
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        phone_id=to_global_id(type="PhoneNode", id=phone.id),
-        phone=phone_data["phone"],
-        phone_type=phone_data["phone_type"],
-        primary=str(phone_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_remove_email(rf, user_gql_client, email_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user, emails=2)
-    primary_email = profile.emails.filter(primary=True).first()
-    email = profile.emails.filter(primary=False).first()
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        removeEmails:[
-                            \"${email_id}\"
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    emails{
-                        edges{
-                        node{
-                            id
-                            email
-                            emailType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "emails": {
-                    "edges": [
-                        {
-                            "node": {
-                                "id": to_global_id(
-                                    type="EmailNode", id=primary_email.id
-                                ),
-                                "email": primary_email.email,
-                                "emailType": primary_email.email_type.name,
-                                "primary": primary_email.primary,
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        email_id=to_global_id(type="EmailNode", id=email.id),
-        email=email_data["email"],
-        email_type=email_data["email_type"],
-        primary=str(email_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_remove_phone(rf, user_gql_client, phone_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    phone = PhoneFactory(profile=profile)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        removePhones:[
-                            \"${phone_id}\"
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    phones{
-                        edges{
-                        node{
-                            id
-                            phone
-                            phoneType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {"updateMyProfile": {"profile": {"phones": {"edges": []}}}}
-
-    mutation = t.substitute(
-        phone_id=to_global_id(type="PhoneNode", id=phone.id),
-        phone=phone_data["phone"],
-        phone_type=phone_data["phone_type"],
-        primary=str(phone_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_remove_address(rf, user_gql_client, address_data):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    address = AddressFactory(profile=profile)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        removeAddresses:[
-                            \"${address_id}\"
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    addresses{
-                        edges{
-                        node{
-                            id
-                            address
-                            addressType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {"updateMyProfile": {"profile": {"addresses": {"edges": []}}}}
-
-    mutation = t.substitute(
-        address_id=to_global_id(type="AddressNode", id=address.id),
-        address=address_data["address"],
-        address_type=address_data["address_type"],
-        primary=str(address_data["primary"]).lower(),
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
 
 
 def test_normal_user_can_query_emails(rf, user_gql_client):
@@ -967,316 +217,6 @@ def test_normal_user_can_query_primary_contact_details(rf, user_gql_client):
         }
     }
     executed = user_gql_client.execute(query, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_change_primary_contact_details(
-    rf, user_gql_client, email_data, phone_data, address_data
-):
-    profile = ProfileFactory(user=user_gql_client.user)
-    PhoneFactory(profile=profile, primary=True)
-    EmailFactory(profile=profile, primary=True)
-    AddressFactory(profile=profile, primary=True)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                            addEmails:[
-                                {emailType: ${email_type}, email:\"${email}\", primary: ${primary}}
-                            ],
-                            addPhones:[
-                                {phoneType: ${phone_type}, phone:\"${phone}\", primary: ${primary}}
-                            ],
-                            addAddresses:[
-                                {
-                                    addressType: ${address_type},
-                                    address:\"${address}\",
-                                    postalCode:\"${postal_code}\",
-                                    city:\"${city}\",
-                                    primary: ${primary}
-                                }
-                            ]
-                        }
-                    }
-                ) {
-                profile{
-                    primaryEmail{
-                        email,
-                        emailType,
-                        primary
-                    },
-                    primaryPhone{
-                        phone,
-                        phoneType,
-                        primary
-                    },
-                    primaryAddress{
-                        address,
-                        postalCode,
-                        city,
-                        addressType,
-                        primary
-                    },
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "primaryEmail": {
-                    "email": email_data["email"],
-                    "emailType": email_data["email_type"],
-                    "primary": True,
-                },
-                "primaryPhone": {
-                    "phone": phone_data["phone"],
-                    "phoneType": phone_data["phone_type"],
-                    "primary": True,
-                },
-                "primaryAddress": {
-                    "address": address_data["address"],
-                    "postalCode": address_data["postal_code"],
-                    "city": address_data["city"],
-                    "addressType": address_data["address_type"],
-                    "primary": True,
-                },
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        email=email_data["email"],
-        email_type=email_data["email_type"],
-        phone=phone_data["phone"],
-        phone_type=phone_data["phone_type"],
-        address=address_data["address"],
-        postal_code=address_data["postal_code"],
-        city=address_data["city"],
-        address_type=address_data["address_type"],
-        primary="true",
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_update_primary_contact_details(
-    rf, user_gql_client, email_data
-):
-    profile = ProfileFactory(user=user_gql_client.user)
-    email = EmailFactory(profile=profile)
-    email_2 = EmailFactory(profile=profile, primary=True)
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                        updateEmails:[
-                            {
-                                id: \"${email_id}\",
-                                emailType: ${email_type},
-                                email:\"${email}\",
-                                primary: ${primary}
-                            }
-                        ]
-                    }
-                }
-            ) {
-                profile{
-                    emails{
-                        edges{
-                        node{
-                            id
-                            email
-                            emailType
-                            primary
-                        }
-                        }
-                    }
-                }
-            }
-            }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "emails": {
-                    "edges": [
-                        {
-                            "node": {
-                                "id": to_global_id(type="EmailNode", id=email.id),
-                                "email": email_data["email"],
-                                "emailType": email_data["email_type"],
-                                "primary": True,
-                            }
-                        },
-                        {
-                            "node": {
-                                "id": to_global_id(type="EmailNode", id=email_2.id),
-                                "email": email_2.email,
-                                "emailType": email_2.email_type.name,
-                                "primary": False,
-                            }
-                        },
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        email_id=to_global_id(type="EmailNode", id=email.id),
-        email=email_data["email"],
-        email_type=email_data["email_type"],
-        primary="true",
-    )
-    executed = user_gql_client.execute(mutation, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_update_sensitive_data(rf, user_gql_client):
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    SensitiveDataFactory(profile=profile, ssn="010199-1234")
-
-    t = Template(
-        """
-            mutation {
-                updateMyProfile(
-                    input: {
-                        profile: {
-                            nickname: "${nickname}"
-                            sensitivedata: {
-                                ssn: "${ssn}"
-                            }
-                        }
-                    }
-                ) {
-                    profile {
-                        nickname
-                        sensitivedata {
-                            ssn
-                        }
-                    }
-                }
-            }
-        """
-    )
-
-    data = {"nickname": "Larry", "ssn": "010199-4321"}
-
-    query = t.substitute(**data)
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "nickname": data["nickname"],
-                "sensitivedata": {"ssn": data["ssn"]},
-            }
-        }
-    }
-    executed = user_gql_client.execute(query, context=request)
-    assert dict(executed["data"]) == expected_data
-
-
-def test_normal_user_can_update_subscriptions_via_profile(rf, user_gql_client):
-    ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    category = SubscriptionTypeCategoryFactory()
-    type_1 = SubscriptionTypeFactory(subscription_type_category=category, code="TEST-1")
-    type_2 = SubscriptionTypeFactory(subscription_type_category=category, code="TEST-2")
-    request = rf.post("/graphql")
-    request.user = user_gql_client.user
-
-    t = Template(
-        """
-        mutation {
-            updateMyProfile(
-                input: {
-                    profile: {
-                        subscriptions: [
-                            {
-                                subscriptionTypeId: \"${type_1_id}\",
-                                enabled: ${type_1_enabled}
-                            },
-                            {
-                                subscriptionTypeId: \"${type_2_id}\",
-                                enabled: ${type_2_enabled}
-                            }
-                        ]
-                    }
-                }
-            ) {
-                profile {
-                    subscriptions {
-                        edges {
-                            node {
-                                enabled
-                                subscriptionType {
-                                    code
-                                    subscriptionTypeCategory {
-                                        code
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        """
-    )
-
-    expected_data = {
-        "updateMyProfile": {
-            "profile": {
-                "subscriptions": {
-                    "edges": [
-                        {
-                            "node": {
-                                "enabled": True,
-                                "subscriptionType": {
-                                    "code": type_1.code,
-                                    "subscriptionTypeCategory": {"code": category.code},
-                                },
-                            }
-                        },
-                        {
-                            "node": {
-                                "enabled": False,
-                                "subscriptionType": {
-                                    "code": type_2.code,
-                                    "subscriptionTypeCategory": {"code": category.code},
-                                },
-                            }
-                        },
-                    ]
-                }
-            }
-        }
-    }
-
-    mutation = t.substitute(
-        type_1_id=to_global_id(type="SubscriptionTypeNode", id=type_1.id),
-        type_1_enabled="true",
-        type_2_id=to_global_id(type="SubscriptionTypeNode", id=type_2.id),
-        type_2_enabled="false",
-    )
-    executed = user_gql_client.execute(mutation, context=request)
     assert dict(executed["data"]) == expected_data
